skip

+++ test/environment/setup.ts
@@ -15,11 +15,14 @@ const API_KEY_USER_ENDPOINT = '/builtin-users/dataverseAdmin/api-token'
 const API_KEY_USER_PASSWORD = 'admin1'
 
 export default async function setupTestEnvironment(): Promise<void> {
-  await setupContainers()
+  await setupContainers() //Set skipContainers to true to skip container setup and run tests against an already running instance
   await setupApiKey()
 }
 
-async function setupContainers(): Promise<void> {
+async function setupContainers(skipContainers?: boolean): Promise<void> {
+  if (skipContainers) {
+    return
+  }
   console.log('Cleaning up old container volumes...')
   fs.rmSync(`${__dirname}/docker-dev-volumes`, { recursive: true, force: true })
   console.log('Running test containers...')
